# inline Javascript into the page