--- a/support-frontend/app/controllers/PayPalRegular.scala
+++ b/support-frontend/app/controllers/PayPalRegular.scala
@@ -68,7 +68,7 @@ class PayPalRegular(
     Ok(views.html.main(
       "Support the Guardian | PayPal Error",
       EmptyDiv("paypal-error-page"),
-      RefPath("payPalErrorPage.js"),
+      Left(RefPath("payPalErrorPage.js")),
       Left(RefPath("payPalErrorPageStyles.css"))
     )()).withSettingsSurrogateKey
   }
@@ -81,7 +81,7 @@ class PayPalRegular(
     Ok(views.html.main(
       "Support the Guardian | PayPal Error",
       EmptyDiv("paypal-error-page"),
-      RefPath("payPalErrorPage.js"),
+      Left(RefPath("payPalErrorPage.js")),
       Left(RefPath("payPalErrorPageStyles.css"))
     )()).withSettingsSurrogateKey
   }