## Why are you doing this?

We are trying to make perf improvements to help people access the page quicker.  This means they are less likely to leave the page before it loads so they are more likely to contribute.

I noticed that with a slow network connection "Fast 3G in chrome", the JS takes 1.9 seconds to download.  Then there is a constant amount of time to get the form on the screen.  See the diagram:
![image](https://user-images.githubusercontent.com/7304387/53893578-60e3fd80-4026-11e9-9ca4-0ed3a959bf91.png)

However if we inline all the JS the connection just gets a clear run from when the main body starts downloading until all the JS is loaded, meaning that there is little contention.
![image](https://user-images.githubusercontent.com/7304387/53893811-dcde4580-4026-11e9-9252-89e72c13c3bb.png)

As you can see the js is fully downloaded 1 second earlier when it's inlined.

Lighthouse difference is small

|before|after|
|---|---|
|![image](https://user-images.githubusercontent.com/7304387/53894027-42323680-4027-11e9-9f95-ef233fe25d67.png)|![image](https://user-images.githubusercontent.com/7304387/53894047-4c543500-4027-11e9-8e8d-883d510a0312.png)|

This is not an AB test, it would be similar difficulty to doing the SSR test, not sure if it's better to just do it, test it, or not do it at all?

The downside could be that people have to download the JS regardless of whether it's been cached.  Although that is flushed out at every deploy anyway, and most people presumably only hit the page once.


[**Trello Card**](https://trello.com/c/8PLMDwNB/358-try-inlining-main-js-bundle-into-html)